game-flow: merge common functions
