+++ src/libtrx/game/game_flow/sequencer.c
@@ -1,0 +1,40 @@
+#include "game/game_flow/sequencer.h"
+
+#include "config.h"
+#include "game/inventory_ring/control.h"
+#include "game/objects/vars.h"
+#include "game/phase.h"
+
+GAME_FLOW_COMMAND GF_EnterPhotoMode(void)
+{
+    PHASE *const subphase = Phase_PhotoMode_Create();
+    const GAME_FLOW_COMMAND gf_cmd = PhaseExecutor_Run(subphase);
+    Phase_PhotoMode_Destroy(subphase);
+    return gf_cmd;
+}
+
+GAME_FLOW_COMMAND GF_PauseGame(void)
+{
+    PHASE *const subphase = Phase_Pause_Create();
+    const GAME_FLOW_COMMAND gf_cmd = PhaseExecutor_Run(subphase);
+    Phase_Pause_Destroy(subphase);
+    return gf_cmd;
+}
+
+GAME_FLOW_COMMAND GF_ShowInventory(const INVENTORY_MODE mode)
+{
+    PHASE *const phase = Phase_Inventory_Create(mode);
+    const GAME_FLOW_COMMAND gf_cmd = PhaseExecutor_Run(phase);
+    Phase_Inventory_Destroy(phase);
+    return gf_cmd;
+}
+
+GAME_FLOW_COMMAND GF_ShowInventoryKeys(const GAME_OBJECT_ID receptacle_type_id)
+{
+    if (g_Config.gameplay.enable_auto_item_selection) {
+        const GAME_OBJECT_ID object_id = Object_GetCognateInverse(
+            receptacle_type_id, g_KeyItemToReceptacleMap);
+        InvRing_SetRequestedObjectID(object_id);
+    }
+    return GF_ShowInventory(INV_KEYS_MODE);
+}
