+++ src/tr1/game/game_flow.c
@@ -1276,16 +1276,6 @@ int32_t GF_GetGymLevelNum(void)
     return g_GameFlow.gym_level_num;
 }
 
-void GF_OverrideCommand(const GAME_FLOW_COMMAND command)
-{
-    g_GameInfo.override_gf_command = command;
-}
-
-GAME_FLOW_COMMAND GF_GetOverrideCommand(void)
-{
-    return g_GameInfo.override_gf_command;
-}
-
 GAME_FLOW_COMMAND GF_PlayAvailableStory(int32_t slot_num)
 {
     GAME_FLOW_COMMAND command = {
@@ -1342,38 +1332,4 @@ GAME_FLOW_COMMAND GF_PlayCutscene(const int32_t level_num)
     const GAME_FLOW_COMMAND gf_cmd = PhaseExecutor_Run(phase);
     Phase_Cutscene_Destroy(phase);
     return gf_cmd;
-}
-
-GAME_FLOW_COMMAND GF_EnterPhotoMode(void)
-{
-    PHASE *const subphase = Phase_PhotoMode_Create();
-    const GAME_FLOW_COMMAND gf_cmd = PhaseExecutor_Run(subphase);
-    Phase_PhotoMode_Destroy(subphase);
-    return gf_cmd;
-}
-
-GAME_FLOW_COMMAND GF_PauseGame(void)
-{
-    PHASE *const subphase = Phase_Pause_Create();
-    const GAME_FLOW_COMMAND gf_cmd = PhaseExecutor_Run(subphase);
-    Phase_Pause_Destroy(subphase);
-    return gf_cmd;
-}
-
-GAME_FLOW_COMMAND GF_ShowInventory(const INVENTORY_MODE mode)
-{
-    PHASE *const phase = Phase_Inventory_Create(mode);
-    const GAME_FLOW_COMMAND gf_cmd = PhaseExecutor_Run(phase);
-    Phase_Inventory_Destroy(phase);
-    return gf_cmd;
-}
-
-GAME_FLOW_COMMAND GF_ShowInventoryKeys(const GAME_OBJECT_ID receptacle_type_id)
-{
-    if (g_Config.gameplay.enable_auto_item_selection) {
-        const GAME_OBJECT_ID object_id = Object_GetCognateInverse(
-            receptacle_type_id, g_KeyItemToReceptacleMap);
-        InvRing_SetRequestedObjectID(object_id);
-    }
-    return GF_ShowInventory(INV_KEYS_MODE);
 }
