+++ src/tr2/game/game_flow/sequencer.c
@@ -5,15 +5,12 @@
 #include "game/demo.h"
 #include "game/fmv.h"
 #include "game/game_flow.h"
-#include "game/inventory_ring.h"
 #include "game/phase.h"
 #include "game/shell.h"
 #include "global/vars.h"
 
-#include <libtrx/config.h>
 #include <libtrx/enum_map.h>
 #include <libtrx/game/game_string_table.h>
-#include <libtrx/game/objects/vars.h>
 #include <libtrx/log.h>
 
 GAME_FLOW_COMMAND GF_DoDemoSequence(int32_t demo_num)
@@ -39,40 +36,6 @@ GAME_FLOW_COMMAND GF_StartGame(
         Phase_Game_Destroy(phase);
         return gf_cmd;
     }
-}
-
-GAME_FLOW_COMMAND GF_EnterPhotoMode(void)
-{
-    PHASE *const subphase = Phase_PhotoMode_Create();
-    const GAME_FLOW_COMMAND gf_cmd = PhaseExecutor_Run(subphase);
-    Phase_PhotoMode_Destroy(subphase);
-    return gf_cmd;
-}
-
-GAME_FLOW_COMMAND GF_PauseGame(void)
-{
-    PHASE *const subphase = Phase_Pause_Create();
-    const GAME_FLOW_COMMAND gf_cmd = PhaseExecutor_Run(subphase);
-    Phase_Pause_Destroy(subphase);
-    return gf_cmd;
-}
-
-GAME_FLOW_COMMAND GF_ShowInventory(const INVENTORY_MODE mode)
-{
-    PHASE *const phase = Phase_Inventory_Create(mode);
-    const GAME_FLOW_COMMAND gf_cmd = PhaseExecutor_Run(phase);
-    Phase_Inventory_Destroy(phase);
-    return gf_cmd;
-}
-
-GAME_FLOW_COMMAND GF_ShowInventoryKeys(const GAME_OBJECT_ID receptacle_type_id)
-{
-    if (g_Config.gameplay.enable_auto_item_selection) {
-        const GAME_OBJECT_ID object_id = Object_GetCognateInverse(
-            receptacle_type_id, g_KeyItemToReceptacleMap);
-        InvRing_SetRequestedObjectID(object_id);
-    }
-    return GF_ShowInventory(INV_KEYS_MODE);
 }
 
 bool GF_DoFrontendSequence(void)
