+++ src/tr2/global/vars.h
@@ -8,8 +8,6 @@
 
 #include <SDL2/SDL.h>
 
-extern GAME_FLOW_COMMAND g_GF_OverrideCommand;
-
 extern int16_t g_RoomsToDraw[MAX_ROOMS_TO_DRAW];
 extern int16_t g_RoomsToDrawCount;
